+++ solidjs-tailwind/src/services/get-gists.js
@@ -1,6 +1,6 @@
 import FetchApi from './api';
 import { useAuth } from '../auth';
-import { GITHUB_GRAPHQL } from '../helper/constants';
+import { GITHUB_GRAPHQL } from '../utils/constants';
 import { USER_GISTS_QUERY } from './queries/gists';
 
 const getGists = async () => {
