+++ solidjs-tailwind/src/services/get-repo-tree.js
@@ -1,6 +1,6 @@
 import FetchApi from './api';
 import { useAuth } from '../auth';
-import { GITHUB_GRAPHQL } from '../helper/constants';
+import { GITHUB_GRAPHQL } from '../utils/constants';
 import { REPO_TREE_QUERY } from './queries/repo-tree';
 
 /**
