+++ solidjs-tailwind/src/services/user-profile.js
@@ -1,6 +1,6 @@
 import FetchApi from './api';
 import { useAuth } from '../auth';
-import { GITHUB_GRAPHQL } from '../helper/constants';
+import { GITHUB_GRAPHQL } from '../utils/constants';
 import { USER_PROFILE_QUERY } from './queries/user-profile';
 
 const userProfile = async (variables) => {
