+++ solidjs-tailwind/src/services/user-repos.js
@@ -1,6 +1,6 @@
 import { useAuth } from '../auth';
 import FetchApi from './api';
-import { GITHUB_GRAPHQL } from '../helper/constants';
+import { GITHUB_GRAPHQL } from '../utils/constants';
 import { USER_REPOS_QUERY } from './queries/all-repos';
 
 const getUserRepos = async (variables) => {
