+++ solidjs-tailwind/src/utils/constants.js
