+++ solidjs-tailwind/src/utils/getFriendlyDate.js
