+++ solidjs-tailwind/src/utils/searchFunctionality.js
