+++ solidjs-tailwind/src/utils/searchFunctionality.spec.js
