+++ solidjs-tailwind/src/utils/typeFilterFunction.js
