+++ solidjs-tailwind/src/utils/typeFilterFunction.spec.js
